ilya -- first commit
Hi Tim,
Here is my first commit to the newly created "ik" branch. There are several goals. The main is to check if my git works and I can commit to my branch. Thus, I check my setup to work on the project.
The changes in this particular test file resolve the commented issue of multiple object definition. The thing is, `test_that()` creates its own environment, and whatever you define in one environment does not communicate with another.  It's all hierarchical.